move the test utils for mls_group to its own directory
to make room for the managed test setup
